requirements.txt
absl-py==2.3.1
accelerate==1.11.0
annotated-types==0.7.0
argbind==0.3.9
asttokens==3.0.0
attrs==25.3.0
audioread==3.1.0
certifi==2025.4.26
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
descript-audiotools @ file:///Users/jaykeecao/Documents/code/SpeeFeaRe/speefeare/audiotools
docstring_parser==0.17.0
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.1
ffmpy==0.6.4
filelock==3.20.0
fire==0.7.1
flatten-dict==0.4.2
fonttools==4.58.1
fsspec==2025.10.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.76.0
h5py==3.15.1
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
julius==0.2.7
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
Markdown==3.10
markdown-it-py==4.0.0
markdown2==2.5.4
MarkupSafe==3.0.3
matplotlib==3.5.0
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.2.3
parso==0.8.5
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
prompt_toolkit==3.0.52
protobuf==3.19.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyloudnorm==0.1.1
PyMatting==1.1.14
pyparsing==3.2.3
pystoi==0.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
randomname==0.2.1
referencing==0.36.2
regex==2025.11.3
rembg==2.0.66
requests==2.32.3
rich==14.2.0
rpds-py==0.25.1
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
sentry-sdk==2.30.0
setproctitle==1.3.6
setuptools-scm==9.2.2
six==1.17.0
smmap==5.0.2
soundfile==0.13.1
soxr==1.0.0
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.2.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tokenizers==0.22.1
tomli==2.3.0
torch==2.9.0
torch-stoi==0.2.3
torchaudio==2.9.0
torchcodec==0.8.1
torchvision==0.24.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
wandb==0.20.1
wcwidth==0.2.14
Werkzeug==3.1.3
